pyproject.toml
[project]
name = "baby-tracker-bot"
version = "2.0.0"
description = "Standalone Multi-tenant Baby Tracker Telegram Bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-telegram-bot>=21.0",
    "sqlalchemy[asyncio]>=2.0.25",
    "alembic>=1.13.1",
    "aiosqlite>=0.20.0",
    "python-dotenv>=1.0.1",
    "stripe>=8.0.0",
    "pydantic-settings>=2.1.0",
    "pysqlcipher3>=1.2.0; sys_platform == 'linux'" 
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.2.1",
    "mypy>=1.8.0",
    "httpx>=0.26.0" 
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
